refactor(TableValidator): interface로 구현하고 구현체를 Order패키지로 이동
