+++ src/main/java/kitchenpos/order/service/TableValidatorImpl.java
@@ -1,0 +1,39 @@
+package kitchenpos.order.service;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import kitchenpos.order.domain.Order.OrderStatus;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.service.TableValidator;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TableValidatorImpl implements TableValidator {
+    private final OrderRepository orderRepository;
+
+    public TableValidatorImpl(OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
+
+    @Override
+    public void validateUpdateEmpty(OrderTable orderTable) {
+        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(
+                orderTable.getId(), Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    @Override
+    public void validateUngroup(List<OrderTable> orderTables) {
+        final List<Long> orderTableIds = orderTables.stream()
+                .map(OrderTable::getId)
+                .collect(Collectors.toList());
+
+        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
+                orderTableIds, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new IllegalArgumentException();
+        }
+    }
+}
