+++ src/main/java/kitchenpos/table/service/TableValidator.java
@@ -1,24 +1,10 @@
 package kitchenpos.table.service;
 
-import java.util.Arrays;
-import kitchenpos.order.domain.Order.OrderStatus;
-import kitchenpos.order.domain.OrderRepository;
+import java.util.List;
 import kitchenpos.table.domain.OrderTable;
-import org.springframework.stereotype.Component;
 
-@Component
-public class TableValidator {
-    private final OrderRepository orderRepository;
+public interface TableValidator {
+    void validateUpdateEmpty(OrderTable orderTable);
 
-    public TableValidator(OrderRepository orderRepository) {
-        this.orderRepository = orderRepository;
-    }
-
-    public void validateUpdateEmpty(OrderTable orderTable) {
-
-        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(
-                orderTable.getId(), Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
-            throw new IllegalArgumentException();
-        }
-    }
+    void validateUngroup(List<OrderTable> orderTables);
 }
